pyproject.toml
[tool.poetry]
name = "fastapi-auth-with-grc"
version = "0.1.0"
description = ""
authors = ["AlekseyLebedik <lebedik.leha@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
grpcio = "^1.62.2"
grpcio-tools = "^1.62.2"
fastapi = "^0.110.1"
python-multipart = "^0.0.9"
uvicorn = "^0.29.0"
requests = "^2.31.0"
protobuf = "4.21.6"
pydantic = {extras = ["email"], version = "^2.7.1"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
asyncpg = "^0.29.0"
pendulum = "^3.0.0"
passlib = "^1.7.4"
bcrypt = "^4.1.2"
annotated-types = "^0.6.0"
anyio = "^4.3.0"
certifi = "^2024.2.2"
charset-normalizer = "^3.3.2"
click = "^8.1.7"
greenlet = "^3.0.3"
h11 = "^0.14.0"
idna = "^3.7"
mako = "^1.3.3"
markdown-it-py = "^3.0.0"
markupsafe = "^2.1.5"
mdurl = "^0.1.2"
pydantic-core = "^2.18.2"
pygments = "^2.17.2"
python-dateutil = "^2.9.0.post0"
python-dotenv = "^1.0.1"
setuptools = "^69.5.1"
six = "^1.16.0"
sniffio = "^1.3.1"
starlette = "^0.37.2"
typing-extensions = "^4.11.0"
tzdata = "^2024.1"
urllib3 = "^2.2.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
loguru = "^0.7.2"
ruff = "^0.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
